Project is now running, but has some bugs
