+++ Views/Shared/_LoginPartial.cshtml
@@ -1,6 +1,6 @@
-@using System.Security.Claims
+@using System.Security.Principal
 
-@if (User.IsSignedIn())
+@if (User.Identity.IsAuthenticated)
 {
     <form asp-controller="Account" asp-action="LogOff" method="post" id="logoutForm" class="navbar-right">
         <ul class="nav navbar-nav navbar-right">
