+++ Views/_ViewImports.cshtml
@@ -2,5 +2,5 @@
 @using codecampster.Models
 @using codecampster.ViewModels.Account
 @using codecampster.ViewModels.Manage
-@using Microsoft.AspNet.Identity
-@addTagHelper "*, Microsoft.AspNet.Mvc.TagHelpers"
+@using Microsoft.AspNetCore.Identity
+@addTagHelper "*, Microsoft.AspNetCore.Mvc.TagHelpers"
